finaltouch

+++ src/components/buttons/PlanBtn.jsx
@@ -10,16 +10,19 @@ const PlanBtn = ({currentPlan}) => {
     const { plan, setPlan } = useContext(WorkOutContext);
 
     const handleSetPlan = () => {
-        console.log('clicked')
-        const newPlanArr = [...plan , currentPlan];
+        const alreadyAdded = plan.find((p) => p.name === currentPlan.name);
+        if (alreadyAdded) {
+            toast.warn(`this plan is already added`);
+            return;
+        }
+        const newPlanArr = [...plan, currentPlan];
         setPlan(newPlanArr);
-    toast.success(`You have read "${currentPlan.name}"`);
-    localStorage.setItem('workout_plan', JSON.stringify(newPlanArr));
-    // alert('success')
+        toast.success(`Added to today's plan!`);
+        localStorage.setItem('workout_plan', JSON.stringify(newPlanArr));
     }
 
     return (
-        <button onClick={()=>handleSetPlan()} className="inline-flex items-center gap-2 bg-[#CCFF00] text-black font-semibold text-sm rounded-lg px-4 py-2.5 ">
+        <button onClick={()=>handleSetPlan()} className="inline-flex items-center gap-2 bg-[#CCFF00] text-black font-semibold text-sm rounded-lg px-4 py-2 ">
               <CalendarPlus size={16} />
               Add to today&apos;s plan
             </button>
